refactor: extract LiveKeyboardOverlayView into focused extensions
Split 1,126-line SwiftUI view into 8 files by responsibility:
- Core (563 lines): state declarations, body, actions
- +ContentBuilders (245): overlayContent, overlayMainContent
- +Styling (166): glassBackground, cursor methods, inspector content
- +ViewModifiers (150): layout handlers, appearance, cursor tracking
- +Header (981): existing, unchanged
- +Inspector (281): existing, unchanged
- +LauncherWelcome (78): existing, unchanged
- +RuleManagement (90): existing, unchanged
